+++ CHANGES.md
@@ -35,6 +35,9 @@ CHANGES
     * Added `AccessTokenDuration` field.
     * Added `RefreshTokenDuration` field.
 
+- `DeviceAuthorizationResponse` struct
+    * Added `Resources` field.
+
 v1.0.5 (2019-09-25)
 -------------------